modified:   Generics/GenericClass/Book.java 	modified:   Generics/GenericClass/Box.java 	modified:   Generics/GenericClass/Fruit.java 	modified:   Generics/GenericClass/TestBoxGenerics.java

+++ Generics/GenericClass/Book.java
@@ -4,9 +4,9 @@ public class Book {
     private String name;
     private String author;
     private int yearOfPublish;
-    private int weight;
+    private double weight;
 
-    public Book(String name, String author, int yearOfPublish, int weight) {
+    public Book(String name, String author, int yearOfPublish, double weight) {
         this.name = name;
         this.author = author;
         this.yearOfPublish = yearOfPublish;
@@ -25,7 +25,7 @@ public class Book {
         return yearOfPublish;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
